+++ pkg/common/constants/constants.go
@@ -33,6 +33,8 @@ const (
 	// Annotations
 	PodGroupAnnotationForPod = "pod-group-name"
 	GpuFraction              = "gpu-fraction"
+	GpuFractionContainerName = "gpu-fraction-container-name"
+	GpuFractionContainerType = "gpu-fraction-container-type"
 	GpuMemory                = "gpu-memory"
 	ReceivedResourceType     = "received-resource-type"
 	GpuFractionsNumDevices   = "gpu-fraction-num-devices"